forgive the long commit... I was in the zone
onDragEnd receives the result prop.

I take the destination object, source and draggableId from result.

destination and source objects have the prop droppableId, the ID of where the draggable object lands.

source is where the object is dragged from and destination is where it lands.

if destination is null, the object returns to the previous state.

the sort action in listActions exports the necessary props to sort the draggable and droppable objects when they are moved, and passes the props to the reducer.

sort is called on the dispatch and the new values for the sort props are given through onDragEnd if there is a valid destination.

in listReducer, a new case is given for CONSTANTS.DRAG_HAPPENED creating a newState with the new props. if the droppable strart and end id is the same, the list where the drag/drop happens is returned as "list".

the cards are spliced on the index where drag happens, grabbing the element. the card is spliced again and inserted where drag ended, with all its props.

the newState is then returned and the drag persists